third!

+++ src/calculateDeposit/DepositType.java
@@ -6,5 +6,5 @@ import java.math.BigDecimal;
  * Created by fateme on 06/08/2016.
  */
 public class DepositType {
-    protected BigDecimal interestRate;
+    protected int interestRate;
 }
